==> clinvar_only/pvs1/preprocessing/8_combine_splice_vep_output.py <==
import pandas as pd


vepDfFull = pd.DataFrame()


for i in range(26):

	part_file_path = f"/net/data/aasubs/clinvar_only/pvs1/{i}_splice_vep_output.vcf"
	vepDfPart = pd.read_csv(part_file_path, sep = "\t", skiprows = 109)

	if i == 0:

		vepDfFull = pd.DataFrame(columns = vepDfPart.columns)

	vepDfFull = pd.concat([vepDfFull, vepDfPart], ignore_index = True)


vepDfFull.to_csv("/net/data/aasubs/clinvar_only/pvs1/splice_vep_output.csv")
